# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  permission.CreatePermissionGroupDto:
    properties:
      description:
        type: string
      name:
        maxLength: 50
        minLength: 5
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - name
    - permissions
    type: object
  permission.PermissionGroupResponseDto:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/permission.PermissionResponseDto'
        type: array
      updated_at:
        type: string
    type: object
  permission.PermissionResponseDto:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  permission.UpdatePermissionGroupDto:
    properties:
      description:
        type: string
      name:
        maxLength: 50
        minLength: 5
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - permissions
    type: object
...
  utils.ErrorResponse:
    properties:
      error:
        example: Error details
        type: string
      message:
        example: Operation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      length:
        type: integer
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@noheir.com
    name: API Support
  description: API server for NoHeir application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NoHeir API
  version: "1.0"
paths:
...
  /permission-groups:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of permission groups with optional filters
      parameters:
      - description: Search term to filter permission groups by name
        in: query
        name: search
        type: string
      - collectionFormat: csv
        description: Filter by permission IDs (UUID format)
        in: query
        items:
          type: string
        name: permissions
        type: array
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Permission groups fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/permission.PermissionGroupResponseDto'
                  type: array
              type: object
        "400":
          description: Invalid pagination parameters or permission IDs
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to fetch permission groups
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get all permission groups
      tags:
      - Permission Groups
    post:
      consumes:
      - application/json
      description: Create a new permission group (role) with specified permissions
      parameters:
      - description: Permission group details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/permission.CreatePermissionGroupDto'
      produces:
      - application/json
      responses:
        "201":
          description: Permission group created successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid request format or invalid permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Role name already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to create permission group
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Create a new permission group
      tags:
      - Permission Groups
  /permission-groups/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing permission group by its UUID
      parameters:
      - description: Permission Group ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Permission group deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Permission group not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to delete permission group or cannot delete Super Admin
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Delete a permission group
      tags:
      - Permission Groups
    get:
      consumes:
      - application/json
      description: Retrieve a single permission group by its UUID
      parameters:
      - description: Permission Group ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Permission group fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/permission.PermissionGroupResponseDto'
              type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Permission group not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to fetch permission group
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get permission group by ID
      tags:
      - Permission Groups
    put:
      consumes:
      - application/json
      description: Update an existing permission group by its UUID
      parameters:
      - description: Permission Group ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: Updated permission group details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/permission.UpdatePermissionGroupDto'
      produces:
      - application/json
      responses:
        "200":
          description: Permission group updated successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid ID format or invalid permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Permission group not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "409":
          description: Role name already exists
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "422":
          description: Validation failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to update permission group
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Update a permission group
      tags:
      - Permission Groups
  /permissions:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all permissions with optional search filter
      parameters:
      - description: Search term to filter permissions by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Permissions fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/permission.PermissionResponseDto'
                  type: array
              type: object
        "403":
          description: Authorization failed
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to fetch permissions
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get all permissions
      tags:
      - Permissions
  /permissions/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single permission by its UUID
      parameters:
      - description: Permission ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Permission fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/permission.PermissionResponseDto'
              type: object
        "400":
          description: Invalid ID format
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "404":
          description: Permission not found
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
        "500":
          description: Failed to fetch permission
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Get permission by ID
      tags:
      - Permissions
...
schemes:
- http
- https
securityDefinitions:
  SessionAuth:
    in: cookie
    name: auth_session
    type: apiKey
swagger: "2.0"